+++ src/main/kotlin/vn/loi/learning/domain/study/memory/model/RatingSource.kt
@@ -2,5 +2,6 @@ package vn.loi.learning.domain.study.memory.model
 
 enum class RatingSource {
     STANDARD_REVIEW,
+    MANUAL_USER,
     MANUAL_USER_OVERRIDE
 }